Add unit tests
This change addresses the need by:

* Testing the ProductTable and AddProductBar components.

Extra:

* Splitting the components into separate files.
